# docs/api.yaml
openapi: 3.0.3
info:
  title: FollowingFeed API
  version: 1.0.0
  description: >-
    FollowingFeed 后端接口契约。除公开接口外，请通过 Bearer Token 认证。
    成功返回 2xx，客户端错误返回 4xx，服务端错误返回 5xx；
    响应体中的 code 用于细分业务结果。除健康检查和文档路由外，业务接口均受
    Redis IP 限流保护：超限返回空响应体的 HTTP 429，限流依赖异常返回空响应体的 HTTP 500。
    需要 Bearer 认证的接口在 Redis 会话检查失败时返回空响应体的 HTTP 503。
servers:
  - url: /api/v1
...
paths:
...
  /users/{id}/follow:
    post:
      tags: [Follow]
      summary: 关注用户
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200": { $ref: "#/components/responses/NoDataResultResponse" }
        "400": { $ref: "#/components/responses/BadRequestResponse" }
        "401": { $ref: "#/components/responses/UnauthorizedResponse" }
        "404": { $ref: "#/components/responses/NotFoundResponse" }
        "409": { $ref: "#/components/responses/ConflictResponse" }
        "500": { $ref: "#/components/responses/InternalServerErrorResponse" }
  /users/{id}/unfollow:
    post:
      tags: [Follow]
      summary: 取消关注用户
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200": { $ref: "#/components/responses/NoDataResultResponse" }
        "400": { $ref: "#/components/responses/BadRequestResponse" }
        "401": { $ref: "#/components/responses/UnauthorizedResponse" }
        "500": { $ref: "#/components/responses/InternalServerErrorResponse" }
  /users/{id}/following/status:
    get:
      tags: [Follow]
      summary: 查询是否关注目标用户
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200": { $ref: "#/components/responses/FollowStatusResultResponse" }
        "400": { $ref: "#/components/responses/BadRequestResponse" }
        "401": { $ref: "#/components/responses/UnauthorizedResponse" }
        "500": { $ref: "#/components/responses/InternalServerErrorResponse" }
  /users/{id}/following:
    get:
      tags: [Follow]
      summary: 获取当前用户的关注列表
      description: id 必须与当前登录用户 ID 一致。
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/UserId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200": { $ref: "#/components/responses/FollowPageResultResponse" }
        "400": { $ref: "#/components/responses/BadRequestResponse" }
        "401": { $ref: "#/components/responses/UnauthorizedResponse" }
        "403": { $ref: "#/components/responses/ForbiddenResponse" }
        "500": { $ref: "#/components/responses/InternalServerErrorResponse" }
  /users/{id}/followers:
    get:
      tags: [Follow]
      summary: 获取当前用户的粉丝列表
      description: id 必须与当前登录用户 ID 一致。
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/UserId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200": { $ref: "#/components/responses/FollowPageResultResponse" }
        "400": { $ref: "#/components/responses/BadRequestResponse" }
        "401": { $ref: "#/components/responses/UnauthorizedResponse" }
        "403": { $ref: "#/components/responses/ForbiddenResponse" }
        "500": { $ref: "#/components/responses/InternalServerErrorResponse" }
...
  /pub/articles/interactions:
    post:
      tags: [Interactive]
      summary: 批量获取公开文章互动统计
      description: 一次最多查询 50 篇文章；没有互动记录的文章返回零值。
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BatchInteractionRequest" }
      responses:
        "200": { $ref: "#/components/responses/BatchInteractionResultResponse" }
        "400": { $ref: "#/components/responses/BadRequestResponse" }
        "413": { $ref: "#/components/responses/PayloadTooLargeResponse" }
        "500": { $ref: "#/components/responses/InternalServerErrorResponse" }
  /pub/articles/{id}/like:
    post:
      tags: [Interactive]
      summary: 点赞文章
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/ArticleId"
      responses:
        "200": { $ref: "#/components/responses/NoDataResultResponse" }
        "400": { $ref: "#/components/responses/BadRequestResponse" }
        "401": { $ref: "#/components/responses/UnauthorizedResponse" }
        "404": { $ref: "#/components/responses/NotFoundResponse" }
        "500": { $ref: "#/components/responses/InternalServerErrorResponse" }
  /pub/articles/{id}/unlike:
    post:
      tags: [Interactive]
      summary: 取消点赞文章
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/ArticleId"
      responses:
        "200": { $ref: "#/components/responses/NoDataResultResponse" }
        "400": { $ref: "#/components/responses/BadRequestResponse" }
        "401": { $ref: "#/components/responses/UnauthorizedResponse" }
        "500": { $ref: "#/components/responses/InternalServerErrorResponse" }
  /pub/articles/{id}/interactions/status:
    get:
      tags: [Interactive]
      summary: 获取文章互动计数和当前用户状态
      description: >-
        匿名访问返回公开计数；携带有效令牌时同时返回当前用户的点赞和收藏状态。
      security:
        - {}
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ArticleId"
      responses:
        "200":
          { $ref: "#/components/responses/InteractionStatusResultResponse" }
        "400": { $ref: "#/components/responses/BadRequestResponse" }
        "401": { $ref: "#/components/responses/UnauthorizedResponse" }
        "500": { $ref: "#/components/responses/InternalServerErrorResponse" }
  /pub/articles/{id}/collect:
    post:
      tags: [Interactive]
      summary: 收藏文章
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/ArticleId"
      responses:
        "200": { $ref: "#/components/responses/NoDataResultResponse" }
        "400": { $ref: "#/components/responses/BadRequestResponse" }
        "401": { $ref: "#/components/responses/UnauthorizedResponse" }
        "404": { $ref: "#/components/responses/NotFoundResponse" }
        "500": { $ref: "#/components/responses/InternalServerErrorResponse" }
  /pub/articles/{id}/uncollect:
    post:
      tags: [Interactive]
      summary: 取消收藏文章
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/ArticleId"
      responses:
        "200": { $ref: "#/components/responses/NoDataResultResponse" }
        "400": { $ref: "#/components/responses/BadRequestResponse" }
        "401": { $ref: "#/components/responses/UnauthorizedResponse" }
        "500": { $ref: "#/components/responses/InternalServerErrorResponse" }
  /pub/articles/{id}/interactions:
    get:
      tags: [Interactive]
      summary: 获取文章公开互动计数
      description: >-
        兼容旧客户端保留；新客户端应使用
        /pub/articles/{id}/interactions/status。
      deprecated: true
      parameters:
        - $ref: "#/components/parameters/ArticleId"
      responses:
        "200": { $ref: "#/components/responses/InteractionCountResultResponse" }
        "400": { $ref: "#/components/responses/BadRequestResponse" }
        "500": { $ref: "#/components/responses/InternalServerErrorResponse" }
  /users/me/collections:
    get:
      tags: [Interactive]
      summary: 获取当前用户的收藏文章
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200": { $ref: "#/components/responses/CollectionPageResultResponse" }
        "400": { $ref: "#/components/responses/BadRequestResponse" }
        "401": { $ref: "#/components/responses/UnauthorizedResponse" }
        "500": { $ref: "#/components/responses/InternalServerErrorResponse" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: followingfeed_refresh_token
  parameters:
    UserId:
      name: id
      in: path
      required: true
      description: 用户 ID
      schema: { type: integer, format: int64, minimum: 1 }
      example: 1
    ArticleId:
      name: id
      in: path
      required: true
      description: 文章 ID
      schema: { type: integer, format: int64, minimum: 1 }
      example: 1
    Page:
      name: page
      in: query
      required: false
      description: 页码，从 1 开始，默认 1
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      required: false
      description: 每页数量，默认 10，最大 50
      schema: { type: integer, minimum: 1, maximum: 50, default: 10 }
  schemas:
...
    BatchInteractionRequest:
      type: object
      required: [ids]
      properties:
        ids:
          type: array
          minItems: 1
          maxItems: 50
          items: { type: integer, format: int64, minimum: 1 }
          example: [1, 2, 3]
    ResultMeta:
      type: object
      required: [code, msg]
      properties:
        code: { type: integer, example: 0 }
        msg: { type: string, example: success }
    NoDataResult:
      allOf:
        - $ref: "#/components/schemas/ResultMeta"
        - type: object
          required: [data]
          properties:
            data: { type: object, nullable: true, example: null }
    BusinessErrorResult:
      type: object
      required: [code, msg, data]
      properties:
        code:
          type: integer
          enum: [4, 5]
          description: 4 表示参数或权限等业务错误，5 表示服务处理失败
          example: 4
        msg: { type: string, example: 参数错误 }
        data: { type: object, nullable: true, example: null }
...
    FollowVO:
      type: object
      required: [id, followerId, followingId, createdAt, updatedAt]
      properties:
        id: { type: integer, format: int64 }
        followerId: { type: integer, format: int64 }
        followingId: { type: integer, format: int64 }
        nickname: { type: string, example: 云端旅人 }
        createdAt: { type: integer, format: int64, description: Unix 毫秒 }
        updatedAt: { type: integer, format: int64, description: Unix 毫秒 }
    FollowPageVO:
      type: object
      required: [items, page, pageSize, total, totalPages]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/FollowVO" }
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1, maximum: 50 }
        total: { type: integer, format: int64, minimum: 0 }
        totalPages: { type: integer, format: int64, minimum: 0 }
    FollowStatusVO:
      type: object
      required: [following]
      properties:
        following: { type: boolean, example: true }
    FollowPageResult:
      allOf:
        - $ref: "#/components/schemas/ResultMeta"
        - type: object
          required: [data]
          properties:
            data: { $ref: "#/components/schemas/FollowPageVO" }
    FollowStatusResult:
      allOf:
        - $ref: "#/components/schemas/ResultMeta"
        - type: object
          required: [data]
          properties:
            data: { $ref: "#/components/schemas/FollowStatusVO" }
    InteractionCountVO:
      type: object
      required: [likeCount, readCount, collectCount]
      properties:
        likeCount: { type: integer, format: int64, minimum: 0 }
        readCount: { type: integer, format: int64, minimum: 0 }
        collectCount: { type: integer, format: int64, minimum: 0 }
    InteractionStatusVO:
      type: object
      required: [liked, likeCount, readCount, collected, collectCount]
      properties:
        liked: { type: boolean }
        likeCount: { type: integer, format: int64, minimum: 0 }
        readCount: { type: integer, format: int64, minimum: 0 }
        collected: { type: boolean }
        collectCount: { type: integer, format: int64, minimum: 0 }
    BatchInteractionData:
      type: object
      description: Key 为文章 ID 字符串，Value 为该文章互动计数
      additionalProperties: { $ref: "#/components/schemas/InteractionCountVO" }
      example:
        "1": { likeCount: 3, readCount: 20, collectCount: 2 }
    InteractionCountResult:
      allOf:
        - $ref: "#/components/schemas/ResultMeta"
        - type: object
          required: [data]
          properties:
            data: { $ref: "#/components/schemas/InteractionCountVO" }
    InteractionStatusResult:
      allOf:
        - $ref: "#/components/schemas/ResultMeta"
        - type: object
          required: [data]
          properties:
            data: { $ref: "#/components/schemas/InteractionStatusVO" }
    BatchInteractionResult:
      allOf:
        - $ref: "#/components/schemas/ResultMeta"
        - type: object
          required: [data]
          properties:
            data: { $ref: "#/components/schemas/BatchInteractionData" }
    CollectedArticleVO:
      type: object
      required: [id, title, abstract, authorId, status, createdAt, updatedAt]
      properties:
        id: { type: integer, format: int64 }
        title: { type: string }
        abstract: { type: string }
        authorId: { type: integer, format: int64 }
        authorNickname: { type: string }
        status:
          {
            type: integer,
            format: int32,
            enum: [2],
            description: 收藏列表仅返回已发布文章,
          }
        createdAt:
          {
            type: integer,
            format: int64,
            description: Unix 毫秒时间戳,
            example: 1786075200000,
          }
        updatedAt:
          {
            type: integer,
            format: int64,
            description: Unix 毫秒时间戳,
            example: 1786075200000,
          }
    CollectionPageVO:
      type: object
      required: [items, page, pageSize, total, totalPages]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/CollectedArticleVO" }
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1, maximum: 50 }
        total: { type: integer, format: int64, minimum: 0 }
        totalPages: { type: integer, format: int64, minimum: 0 }
    CollectionPageResult:
      allOf:
        - $ref: "#/components/schemas/ResultMeta"
        - type: object
          required: [data]
          properties:
            data: { $ref: "#/components/schemas/CollectionPageVO" }
  responses:
...
    NoDataResultResponse:
      description: 操作成功时 data 为 null
      content:
        application/json:
          schema:
            oneOf:
              - $ref: "#/components/schemas/NoDataResult"
              - $ref: "#/components/schemas/BusinessErrorResult"
...
    FollowPageResultResponse:
      description: 关注或粉丝分页结果
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/FollowPageResult" } },
        }
    FollowStatusResultResponse:
      description: 当前用户对目标用户的关注状态
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/FollowStatusResult" } },
        }
    InteractionCountResultResponse:
      description: 文章公开互动计数
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/InteractionCountResult" } },
        }
    InteractionStatusResultResponse:
      description: 当前用户的文章点赞和收藏状态
      content:
        {
          application/json:
            {
              schema: { $ref: "#/components/schemas/InteractionStatusResult" },
            },
        }
    BatchInteractionResultResponse:
      description: 多篇文章的公开互动计数
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/BatchInteractionResult" } },
        }
    CollectionPageResultResponse:
      description: 当前用户收藏文章分页结果
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/CollectionPageResult" } },
        }
...
    BadRequestResponse:
      description: 请求格式或参数错误
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/BusinessErrorResult" } },
        }
    PayloadTooLargeResponse:
      description: 请求体超过 1 MiB 限制
      content:
        application/json:
          schema: { $ref: "#/components/schemas/BusinessErrorResult" }
    ForbiddenResponse:
      description: 已认证，但没有资源访问权限
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/BusinessErrorResult" } },
        }
    NotFoundResponse:
      description: 请求的资源不存在
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/BusinessErrorResult" } },
        }
    ConflictResponse:
      description: 请求与资源当前状态冲突
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/BusinessErrorResult" } },
        }
    InternalServerErrorResponse:
      description: 服务内部错误
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/BusinessErrorResult" } },
        }
    UnauthorizedResponse:
      description: Token 缺失、无效或会话已失效；响应体为空